MIM-216 删除不可达的旧 SessionRuntime
