+++ internal/httpapi/router_endpoints_test.go
@@ -2,7 +2,6 @@ package httpapi
 
 import (
 	"bytes"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -586,65 +585,6 @@ func canonicalTestPath(t *testing.T, path string) string {
 	return realPath
 }
 
-func TestActiveSessionSnapshotsFiltersByProjectBeforePagination(t *testing.T) {
-	now := time.Unix(100, 0)
-	list := []*session.Session{
-		{ID: "sess_demo", ProjectID: "demo", Title: "demo", Status: "running", UpdatedAt: now},
-		{ID: "sess_other", ProjectID: "other", Title: "other", Status: "running", UpdatedAt: now},
-	}
-
-	all := activeSessionSnapshots(list, "")
-	if len(all) != 2 {
-		t.Fatalf("全局列表应保留所有运行会话，got=%v", all)
-	}
-
-	demo := activeSessionSnapshots(list, "demo")
-	if len(demo) != 1 || demo[0].ID != "sess_demo" {
-		t.Fatalf("项目列表应在 snapshot 阶段排除无关运行会话，got=%v", demo)
-	}
-
-	missing := activeSessionSnapshots(list, "missing")
-	if len(missing) != 0 {
-		t.Fatalf("未知项目不应保留运行会话，got=%v", missing)
-	}
-}
-
-func TestActiveSessionSnapshotWindowUsesCursorAndBoundedTopK(t *testing.T) {
-	now := time.UnixMilli(1_780_308_003_000)
-	list := []*session.Session{
-		{ID: "sess_alpha", ProjectID: "demo", Title: "alpha", Status: "running", UpdatedAt: now},
-		{ID: "sess_delta", ProjectID: "demo", Title: "delta", Status: "running", UpdatedAt: now},
-		{ID: "sess_beta", ProjectID: "demo", Title: "beta", Status: "running", UpdatedAt: now},
-		{ID: "sess_gamma", ProjectID: "demo", Title: "gamma", Status: "running", UpdatedAt: now},
-		{ID: "sess_other", ProjectID: "other", Title: "other", Status: "running", UpdatedAt: now.Add(time.Second)},
-	}
-
-	firstWindow := activeSessionSnapshotWindow(list, "demo", sessionPageCursor{}, false, 2)
-	if got := sessionSnapshotIDs(firstWindow); len(got) != 2 || got[0] != "sess_gamma" || got[1] != "sess_delta" {
-		t.Fatalf("active window 应只保留按 updated_at/id 排序后的 top K，got=%v", got)
-	}
-
-	cursor := sessionPageCursor{ID: "sess_delta", UpdatedAtMS: now.UnixMilli()}
-	secondWindow := activeSessionSnapshotWindow(list, "demo", cursor, true, 2)
-	if got := sessionSnapshotIDs(secondWindow); len(got) != 2 || got[0] != "sess_beta" || got[1] != "sess_alpha" {
-		t.Fatalf("active window 应在 cursor 后继续并保持稳定 id tie-breaker，got=%v", got)
-	}
-}
-
-func TestDecodeSessionCursorRejectsMalformedNonEmptyCursor(t *testing.T) {
-	if _, hasCursor, err := decodeSessionCursor(""); err != nil || hasCursor {
-		t.Fatalf("空 cursor 应被视为未分页，has=%v err=%v", hasCursor, err)
-	}
-
-	invalidJSON := base64.RawURLEncoding.EncodeToString([]byte("{"))
-	missingFields := base64.RawURLEncoding.EncodeToString([]byte(`{"id":"sess_1"}`))
-	for _, raw := range []string{"not-base64!", invalidJSON, missingFields} {
-		if _, hasCursor, err := decodeSessionCursor(raw); err == nil || hasCursor {
-			t.Fatalf("非空无效 cursor 应返回错误且不启用分页，raw=%q has=%v err=%v", raw, hasCursor, err)
-		}
-	}
-}
-
 func authedRequest(t *testing.T, method, path string, body any) *http.Request {
 	t.Helper()
 
@@ -710,14 +650,6 @@ func containsWorktreeBranch(items []worktreeBranchItem, name string, kind string
 		}
 	}
 	return false
-}
-
-func sessionSnapshotIDs(items []session.SessionSnapshot) []string {
-	ids := make([]string, 0, len(items))
-	for _, item := range items {
-		ids = append(ids, item.ID)
-	}
-	return ids
 }
 
 func TestHealthzDoesNotRequireAuth(t *testing.T) {
